test_api_apoena/pyproject.toml
[tool.poetry]
name = "test-api-aponea"
version = "0.1.0"
description = ""
authors = ["iuryrosal <iuryrosal@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-auth = "^2.29.0"
google-crc32c = "^1.5.0"
requests = "^2.32.3"
google-cloud-secret-manager = "^2.20.0"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
